fix: consistent lists round-trip
