+++ packages/app/src/features/NoteEditor/RichEditor/plugins/Markdown/markdownParser.ts
@@ -4,74 +4,34 @@ import {
 	$createParagraphNode,
 	$createTextNode,
 	$getRoot,
-	$isLineBreakNode,
-	$isParagraphNode,
 	$isTextNode,
 	IS_CODE,
 	LexicalNode,
 } from 'lexical';
-import {
-	Blockquote,
-	Break,
-	Code,
-	Content,
-	Delete,
-	Emphasis,
-	Heading,
-	HTML,
-	Image,
-	InlineCode,
-	Link,
-	List,
-	ListItem,
-	Paragraph,
-	Root,
-	Strong,
-	Table,
-	TableCell,
-	TableRow,
-	Text,
-	ThematicBreak,
-} from 'mdast';
+import { Content, Root } from 'mdast';
 import remarkGfm from 'remark-gfm';
 import remarkParse from 'remark-parse';
 import remarkStringify from 'remark-stringify';
 import { unified } from 'unified';
 import { u } from 'unist-builder';
-import { $createCodeNode, $isCodeNode } from '@lexical/code';
-import { $createLinkNode, $isLinkNode } from '@lexical/link';
-import {
-	$createListItemNode,
-	$createListNode,
-	$isListItemNode,
-	$isListNode,
-	ListType,
-} from '@lexical/list';
-import {
-	$createHorizontalRuleNode,
-	$isHorizontalRuleNode,
-} from '@lexical/react/LexicalHorizontalRuleNode';
-import {
-	$createHeadingNode,
-	$createQuoteNode,
-	$isHeadingNode,
-	$isQuoteNode,
-} from '@lexical/rich-text';
+import { $createCodeNode } from '@lexical/code';
+import { $createLinkNode } from '@lexical/link';
+import { $createListItemNode, $createListNode, ListType } from '@lexical/list';
+import { $createHorizontalRuleNode } from '@lexical/react/LexicalHorizontalRuleNode';
+import { $createHeadingNode, $createQuoteNode } from '@lexical/rich-text';
 import {
 	$createTableCellNode,
 	$createTableNode,
 	$createTableRowNode,
-	$isTableCellNode,
-	$isTableNode,
-	$isTableRowNode,
 	TableCellHeaderStates,
 } from '@lexical/table';
 
-import { $createImageNode, $isImageNode } from '../Image/ImageNode';
-import { $createFormattingNode, $isFormattingNode } from './nodes/FormattingNode';
+import { $createImageNode } from '../Image/ImageNode';
+import { convertLexicalNodeToMarkdownNode } from './convertLexicalNodeToMarkdownNode';
+import { $createFormattingNode } from './nodes/FormattingNode';
 import { $createRawNode } from './nodes/RawNode';
 
-const markdownProcessor = unified()
+export const markdownProcessor = unified()
 	.use(remarkParse)
 	.use(remarkGfm)
 	.use(remarkStringify, {
@@ -269,172 +229,9 @@ export const $convertFromMarkdownString = (rawMarkdown: string) => {
 
 export const $convertToMarkdownString = () => {
 	const rootNode = $getRoot();
-
-	const convertToMarkdownNode = (node: LexicalNode): Content => {
-		if ($isParagraphNode(node)) {
-			const paragraph = u('paragraph', { children: [] }) as Paragraph;
-
-			const nestedNodes = node.getChildren();
-			if (
-				nestedNodes.length === 0 ||
-				(nestedNodes.every((node) => $isTextNode(node)) &&
-					node.getTextContent().trim().length === 0)
-			) {
-				paragraph.children.push(u('text', { value: '' }) satisfies Text);
-				return paragraph;
-			}
-
-			for (const child of node.getChildren()) {
-				const content = convertToMarkdownNode(child);
-				paragraph.children.push(content as any);
-			}
-
-			return paragraph;
-		}
-
-		if ($isImageNode(node)) {
-			return u('image', {
-				url: node.getSrc(),
-				alt: node.getAltText(),
-			}) satisfies Image;
-		}
-
-		if ($isTextNode(node)) {
-			if (node.hasFormat('code')) {
-				return u('inlineCode', {
-					value: node.getTextContent(),
-				}) satisfies InlineCode;
-			}
-
-			return u('text', { value: node.getTextContent() }) satisfies Text;
-		}
-
-		if ($isCodeNode(node)) {
-			return u('code', {
-				lang: node.getLanguage(),
-				value: node.getTextContent(),
-			}) satisfies Code;
-		}
-
-		if ($isFormattingNode(node)) {
-			const tagName = node.getTagName();
-			switch (tagName) {
-				case 'em': {
-					return u('emphasis', {
-						children: node
-							.getChildren()
-							.map(convertToMarkdownNode) as Emphasis['children'],
-					}) satisfies Emphasis;
-				}
-				case 'del': {
-					return u('delete', {
-						children: node
-							.getChildren()
-							.map(convertToMarkdownNode) as Delete['children'],
-					}) satisfies Delete;
-				}
-				case 'b': {
-					return u('strong', {
-						children: node
-							.getChildren()
-							.map(convertToMarkdownNode) as Strong['children'],
-					}) satisfies Strong;
-				}
-			}
-		}
-
-		if ($isHorizontalRuleNode(node)) {
-			return u('thematicBreak') satisfies ThematicBreak;
-		}
-
-		if ($isListNode(node)) {
-			return u('list', {
-				ordered: node.getTag() === 'ol',
-				start: node.getStart(),
-				spread: false,
-				children: node
-					.getChildren()
-					.map(convertToMarkdownNode) as List['children'],
-			}) satisfies List;
-		}
-		if ($isListItemNode(node)) {
-			return u('listItem', {
-				spread: false,
-				checked: node.getChecked(),
-				children: [
-					u('paragraph', {
-						children: node
-							.getChildren()
-							.map(convertToMarkdownNode) as Paragraph['children'],
-					}) as Paragraph,
-				],
-			}) satisfies ListItem;
-		}
-
-		if ($isLinkNode(node)) {
-			return u('link', {
-				url: node.getURL(),
-				alt: node.getTitle(),
-				children: node
-					.getChildren()
-					.map(convertToMarkdownNode) as Link['children'],
-			}) satisfies Link;
-		}
-
-		if ($isQuoteNode(node)) {
-			return u('blockquote', {
-				children: node
-					.getChildren()
-					.map(convertToMarkdownNode) as Blockquote['children'],
-			}) satisfies Blockquote;
-		}
-
-		if ($isTableNode(node)) {
-			return u('table', {
-				children: node
-					.getChildren()
-					.map(convertToMarkdownNode) as Table['children'],
-			}) satisfies Table;
-		}
-		if ($isTableRowNode(node)) {
-			return u('tableRow', {
-				children: node
-					.getChildren()
-					.map(convertToMarkdownNode) as TableRow['children'],
-			}) satisfies TableRow;
-		}
-		if ($isTableCellNode(node)) {
-			return u('tableCell', {
-				children: node
-					.getChildren()
-					.map(convertToMarkdownNode) as TableCell['children'],
-			}) satisfies TableCell;
-		}
-
-		if ($isLineBreakNode(node)) {
-			return u('break', {}) satisfies Break;
-		}
-
-		if ($isHeadingNode(node)) {
-			const depth = Math.max(
-				1,
-				Math.min(parseInt(node.getTag().slice(1)), 6),
-			) as Heading['depth'];
-			return u('heading', {
-				depth: depth,
-				children: node
-					.getChildren()
-					.map(convertToMarkdownNode) as Heading['children'],
-			}) satisfies Heading;
-		}
-
-		// Default node
-		return u('html', { value: node.getTextContent() }) as HTML;
-	};
-
 	const children = rootNode.getChildren();
 	const mdTree = u('root', {
-		children: children.map(convertToMarkdownNode),
+		children: children.map(convertLexicalNodeToMarkdownNode),
 	}) satisfies Root;
 
 	return markdownProcessor.stringify(mdTree);
